+++ src/plone/testing/z2.py
@@ -7,12 +7,13 @@ from plone.testing import zca
 from plone.testing import zodb
 from plone.testing._z2_testbrowser import Browser  # noqa # BBB
 from Testing.ZopeTestCase.ZopeLite import _patched as ZOPETESTCASEALERT
+from Zope2.App.schema import Zope2VocabularyRegistry
 from zope.schema.vocabulary import getVocabularyRegistry
 from zope.schema.vocabulary import setVocabularyRegistry
-from Zope2.App.schema import Zope2VocabularyRegistry
 
 import contextlib
 import os
+import transaction
 
 
 _INSTALLED_PRODUCTS = {}
@@ -247,9 +248,7 @@ def zopeApp(db=None, connection=None, environ=None):
     pass an open connection as ``connection`` (the connection will not be
     closed).
     """
-
     import Zope2
-    import transaction
 
     closeConn = True
     if connection is not None:
@@ -286,10 +285,10 @@ def zopeApp(db=None, connection=None, environ=None):
     finally:
         try:
             app.REQUEST.close()
-
             if closeConn:
+                transaction.abort()
                 connection.close()
-        except:
+        except Exception:
             if inner_exception:
                 raise inner_exception
             else:
@@ -591,6 +590,7 @@ class Startup(Layer):
         del self._dbtab
 
         # Close and pop the zodbDB resource
+        transaction.abort()
         self['zodbDB'].close()
         del self['zodbDB']
 
@@ -740,6 +740,7 @@ class Startup(Layer):
 
         setVocabularyRegistry(self._oldVocabularyRegistry)
 
+
 STARTUP = Startup()
 
 
@@ -776,7 +777,6 @@ class IntegrationTesting(Layer):
 
     def testSetUp(self):
         import Zope2
-        import transaction
 
         # Open a new app and save it as the resource ``app``.
 
@@ -813,7 +813,6 @@ here and let you fix it.""")
         self['request'] = request
 
     def testTearDown(self):
-        import transaction
 
         # Abort the transaction
         transaction.abort()
@@ -869,7 +868,6 @@ class FunctionalTesting(Layer):
 
     def testSetUp(self):
         import Zope2
-        import transaction
 
         # Override zodbDB from the layer setup. Since it was set up by
         # this layer, we can't just assign a new shadow. We therefore keep
@@ -905,8 +903,6 @@ class FunctionalTesting(Layer):
         self['request'] = request
 
     def testTearDown(self):
-        import transaction
-
         # Abort any open transactions
         transaction.abort()