+++ CHANGES.rst
@@ -13,6 +13,9 @@ Changelog
 - Fix ``.z2.Startup.setUpZCML()`` to be compatible with Zope >= 4.0a2.
   [icemac]
 
+- Fix version pins on the package itself to be able to run the tests.
+  [gforcada]
+
 5.0.0 (2016-02-19)
 ------------------